requirements.txt
Django
django-bootstrap3
django-location-field
django-geoposition
django-crispy-forms
